moved utils.py and spec.py
