+++ dictos/series/taylor_expansion.py
@@ -1,6 +1,6 @@
 import sympy as sp
 
-from dictos.spec import DEFAULT_DIFFERENTIAND, is_not_natural_number
+from dictos.utilities.spec import DEFAULT_DIFFERENTIAND, is_not_natural_number
 from dictos.linalg.linalg import dot_product
 from dictos.error.taylor_expansion import (
     UnsupportedOrderOfDerivativeError,
