+++ dictos/utilities/spec.py
